registrar/requirements.txt
Flask
python-gvm
pyyaml